requirements.txt
flask
openai
python-dotenv
google-cloud-texttospeech
requests